GameGate优化

+++ src/GameGate/Package/TSessionInfo.cs
@@ -21,13 +21,13 @@ namespace GameGate
         public int nCheckSendLength;
         public bool boSendAvailable;
         public bool boSendCheck;
-        public long dwTimeOutTime;
-        public long dwReceiveTick;
+        public int dwTimeOutTime;
+        public int dwReceiveTick;
         public string sRemoteAddr;
         /// <summary>
         /// 发言间隔控制
         /// </summary>
-        public long dwSayMsgTick;
+        public int dwSayMsgTick;
         /// <summary>
         /// 开启用户延迟处理
         /// </summary>
@@ -35,7 +35,7 @@ namespace GameGate
         /// <summary>
         /// 延迟发送给客户端间隔
         /// </summary>
-        public long dwClientCheckTimeOut;
+        public int dwClientCheckTimeOut;
         public string sUserName;
     }
 }
